feat(core): 식량 소모를 인게임 경과시간 기준으로 (Framework 배율 연동)

+++ Assets/_Project/11.CoreServices/Scripts/Save/SaveData.cs
@@ -242,7 +242,7 @@ namespace ND.Framework
         public string itemId = string.Empty;
         public string itemName = string.Empty;
         public float weight;
-        public int basePrice;
+        public long basePrice;
     }
 
     /// <summary>
